+++ h3ron/src/experimental.rs
@@ -82,8 +82,8 @@ mod tests {
     fn test_local_ij() {
         let origin_cell = H3Cell::try_from(0x89283080ddbffff_u64).unwrap();
         let ring = origin_cell.k_ring(1);
-        assert_ne!(ring.len(), 0);
-        let other_cell = *ring.iter().find(|i| **i != origin_cell).unwrap();
+        assert_ne!(ring.iter().count(), 0);
+        let other_cell = ring.iter().find(|i| *i != origin_cell).unwrap();
 
         // the coordij of the origin index. This is not necessarily at (0, 0)
         let coordij_origin = h3_to_local_ij(&origin_cell, &origin_cell).unwrap();